perf: add support for combined event/stack lines
It turns out that sometimes, `perf script` spits out event lines that
_also_ include stack information, like so:

    false 64414 20110.539243:       1893 cycles:u:      7fd5bd78f145 _dl_start+0x3c5 (/usr/lib/ld-2.31.so)

Which we would not handle correctly. This patch adds support for those
types of lines. I am not aware of anyone having run into this, but it
can happen if you do something like `perf record /bin/false`.